Fixes the problem with HLS.
When m3u8 field is provided in json, request to http stream will result
in a 404 Not found. That means either HLS or HTTP is available.
`
Input #0, mpegts, from 'pipe:0':
  Duration: N/A, start: 40169.165733, bitrate: 60 kb/s
  Program 1
    Metadata:
      service_name    : Service01
      service_provider: FFmpeg
    Stream #0.0[0x100]: Video: h264 (High), yuv420p, 504x896, 15 fps,
90k tbn
    Stream #0.1[0x101]: Audio: aac, 44100 Hz, mono, fltp, 60 kb/s
`
`
Input #0, flv, from 'pipe:0':
  Metadata:
    copyright       : lvll
    hasVideo        : true
    hasAudio        : true
    displayWidth    : 504
    displayHeight   : 896
  Duration: N/A, start: 0.000000, bitrate: N/A
    Stream #0.0: Video: h264 (High), yuv420p, 504x896, 14.42 fps, 1k tbn
    Stream #0.1: Audio: aac, 44100 Hz, mono, fltp
`
HLS and HTTP stream seems to have different quality, HLS is better than
HTTP. But either is available, so `STREAM_WEIGHTS` not set and only an
`live` option provided.